pyproject.toml
[project]
name = "app-laundry"
version = "0.0.0"
description = "Laundry management web app"
requires-python = ">=3.12"

dependencies = [
	"alembic==1.13.1",
	"bidict==0.23.1",
	"black==25.1.0",
	"blinker==1.9.0",
	"certifi==2025.8.3",
	"charset-normalizer==3.4.3",
	"click==8.2.1",
	"dnspython==2.7.0",
	"email_validator==2.1.1",
	"flake8==6.1.0",
	"Flask==3.0.3",
	"Flask-Login==0.6.3",
	"Flask-Mail==0.9.1",
	"Flask-SocketIO==5.5.1",
	"Flask-SQLAlchemy==3.1.1",
	"greenlet==3.2.4",
	"h11==0.16.0",
	"idna==3.10",
	"isort==6.0.1",
	"itsdangerous==2.2.0",
	"Jinja2==3.1.6",
	"Mako==1.3.10",
	"MarkupSafe==3.0.2",
	"mccabe==0.7.0",
	"mypy_extensions==1.1.0",
	"packaging==25.0",
	"pathspec==0.12.1",
	"platformdirs==4.3.8",
	"pycodestyle==2.11.1",
	"pyflakes==3.1.0",
	"python-dotenv==1.0.1",
	"python-engineio==4.12.2",
	"python-socketio==5.13.0",
	"requests==2.31.0",
	"ruff==0.12.10",
	"simple-websocket==1.1.0",
	"SQLAlchemy==2.0.43",
	"typing_extensions==4.14.1",
	"urllib3==2.5.0",
	"Werkzeug==3.0.3",
	"wsproto==1.2.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
